# Optimize raycasting with max distance checking