--- a/renders/renders.md
+++ b/renders/renders.md
@@ -4,43 +4,43 @@ _spp = samples per pixel_
 
 #### `scenes/scene.json`
 
-800 x 800 pixels (4 spp), 1,524 primitives, 112,811,095 rays, 34.393s on i7 8650U
+800 x 800 pixels (4 spp), 1,524 primitives, 112,806,820 rays, 29.305s on i7 8650U
 
 ![scene.json](./scene.png)
 
 #### `scenes/mesh.json`
 
-800 x 800 pixels (8 spp), 2,269 primitives, 70,711,320 rays, 22.744s on i7 8650U
+800 x 800 pixels (8 spp), 2,269 primitives, 70,710,256 rays, 10.766s on i7 8650U
 
 ![mesh.json](./mesh.png)
 
 #### `scenes/refraction.json`
 
-800 x 800 pixels (4 spp), 6 primitives, 188,986,845 rays, 12.274s on i7 8650U
+800 x 800 pixels (4 spp), 6 primitives, 188,994,416 rays, 11.754s on i7 8650U
 
 ![refraction.json](./refraction.png)
 
 #### `scenes/physical.json`
 
-800 x 800 pixels (4 spp), 50 primitives, 453,358,438 rays, 47.895s on i7 8650U
+800 x 800 pixels (4 spp), 50 primitives, 453,360,287 rays, 43.228s on i7 8650U
 
 ![physical.json](./physical.png)
 
 #### `scenes/mesh_test.json`
 
-800 x 800 pixels (4 spp), 41,372 primitives, 39,250,881 rays, 23.666s on i7 8650U
+800 x 800 pixels (4 spp), 41,372 primitives, 39,251,578 rays, 8.108s on i7 8650U
 
 ![mesh_test.json](./mesh_test.png)
 
 #### `scenes/occlusion.json`
 
-800 x 800 pixels (4 spp), 8 primitives, 29,345,106 rays, 2.141s on i7 8650U
+800 x 800 pixels (4 spp), 8 primitives, 29,345,257 rays, 1.929s on i7 8650U
 
 ![occlusion.json](./occlusion.png)
 
 #### `scenes/test.json`
 
-800 x 800 pixels (4 spp), 9 primitives, 35,717,574 rays, 2.975s on i7 8650U
+800 x 800 pixels (4 spp), 9 primitives, 35,717,895 rays, 2.683s on i7 8650U
 
 ![test.json](./test.png)