--- a/src/primitives/triangle.rs
+++ b/src/primitives/triangle.rs
@@ -191,7 +191,7 @@ impl Transformed for RaytracingTriangle {
 }
 
 impl Intersectable for RaytracingTriangle {
-    fn intersect(&self, ray: &Ray) -> Option<Intersection> {
+    fn intersect(&self, ray: &Ray, max_distance: Option<f64>) -> Option<Intersection> {
         let edge1 = self.vertex_data[1].position - self.vertex_data[0].position;
         let edge2 = self.vertex_data[2].position - self.vertex_data[0].position;
         let p_vec = ray.direction.cross(&edge2);
@@ -218,6 +218,10 @@ impl Intersectable for RaytracingTriangle {
         }
 
         let distance = edge2.dot(&q_vec) / det;
+
+        if distance < 0.0 || (max_distance.is_some() && max_distance.unwrap() < distance) {
+            return None;
+        }
 
         Some(Intersection::new_with_data(
             self,