… renders/mesh.png
… renders/mesh_test.png
… renders/occlusion.png
… renders/physical.png
… renders/refraction.png
… renders/scene.png
… renders/test.png