Currently, every raycast runs an intersection test on every primitive whose bounding box intersects the ray and then sorts by distance to return the closest intersection. This optimization orders the nodes to traverse in order along the ray and passes the closest intersection distance along to terminate early on nodes that are past the closest intersection.

Benchmarks on an i7 8650U (8 cores, averaged over 3 runs):
|scene|unoptimized|optimized|render time change|
|-|-|-|-|
|`scene.json` (1,524 primitives)|34.393s|29.305s|1.17x faster|
|`mesh.json` (2,269 primitives)|22.744s|10.766s|2.11x faster|
|`refraction.json` (6 primitives)|12.274s|11.754s|1.04x faster|
|`physical.json` (50 primitives)|47.895s|43.228s|1.11x faster|
|`mesh_test.json` (41,372 primitives)|23.666s|8.108s|2.92x faster|
|`occlusion.json` (8 primitives)|2.141s|1.929s|1.11x faster|
|`test.json` (9 primitives)|2.975s|2.683s|1.11x faster|